Interview/app/requirements.txt
vllm==0.8.4
transformers==4.52.3
torch==2.6.0
numpy==2.2.6
sentence-transformers==4.1.0
qdrant-client==1.14.2
huggingface-hub==0.30.2
langchain==0.3.25
langgraph==0.4.0
langchain-core==0.3.61
langsmith==0.3.42
pydantic==2.11.5
fastapi==0.115.12
uvicorn==0.34.2
tenacity==9.1.2
python-dotenv==1.1.0
tavily-python==0.7.3
pymysql==1.1.1
evaluate==0.4.3
bert-score==0.3.13
rouge-score==0.1.2
absl-py==2.3.0
nltk==3.9.1
